# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repair-requests:
    post:
      operationId: createRepairRequest
      tags: [repair-requests]
      summary: Submit a repair request
      description: Stores an anonymous repair request and sends business and customer email notifications.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairRequestInput"
      responses:
        "201":
          description: Repair request accepted and both emails sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairRequestSubmission"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Database or server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Email delivery failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/repair-requests:
    get:
      operationId: getAdminRepairRequests
      tags: [admin]
      summary: List all contact and repair requests
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AdminRequestStatus"
        - name: customerType
          in: query
          required: false
          schema:
            type: string
            enum: [guest, registered]
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        "200":
          description: All contact and repair requests visible to administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminRepairRequest"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/repair-requests/{requestId}:
    get:
      operationId: getAdminRepairRequest
      tags: [admin]
      summary: View a complete repair request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complete repair request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRepairRequest"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Repair request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAdminRepairRequest
      tags: [admin]
      summary: Update a repair request status or internal notes
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRepairRequestUpdate"
      responses:
        "200":
          description: Repair request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRepairRequest"
        "400":
          description: Invalid update data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Repair request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/bookings:
    get:
      operationId: getAdminBookings
      tags: [admin]
      summary: List all customer bookings
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AdminBookingStatus"
        - name: customerType
          in: query
          required: false
          schema:
            type: string
            enum: [guest, registered]
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        "200":
          description: All customer bookings visible to administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBooking"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/bookings/{bookingId}:
    get:
      operationId: getAdminBooking
      tags: [admin]
      summary: View a complete customer booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complete booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBooking"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAdminBooking
      tags: [admin]
      summary: Update a booking status or internal notes
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBookingUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBooking"
        "400":
          description: Invalid update data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Administrator authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    RepairRequestInput:
      type: object
      additionalProperties: false
      required:
        - customerName
        - phone
        - email
        - applianceType
        - problemDescription
        - address
      properties:
        customerName:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type: string
          pattern: "^[0-9]{10}$"
        email:
          type: string
          format: email
          maxLength: 254
        applianceType:
          type: string
          enum:
            - Washing Machine Repair
            - Refrigerator Repair
            - Micro Oven Repair
            - LED TV Repair
        problemDescription:
          type: string
          minLength: 10
          maxLength: 2000
        address:
          type: string
          minLength: 10
          maxLength: 500
        preferredDate:
          type: ["string", "null"]
          format: date
        preferredTime:
          type: ["string", "null"]
          maxLength: 100
    RepairRequestSubmission:
      type: object
      required:
        - success
        - requestId
        - message
      properties:
        success:
          type: boolean
        requestId:
          type: string
        message:
          type: string
...
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    AdminRequestStatus:
      type: string
      enum:
        - pending
        - contacted
        - in_progress
        - completed
        - cancelled
    AdminRepairRequest:
      type: object
      required:
        - requestId
        - customerName
        - email
        - phone
        - applianceType
        - problemDescription
        - address
        - customerId
        - customerType
        - preferredDate
        - preferredTime
        - status
        - adminNotes
        - cancellationReason
        - emailStatus
        - createdAt
        - updatedAt
      properties:
        requestId:
          type: string
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        applianceType:
          type: string
        problemDescription:
          type: string
        address:
          type: string
        customerId:
          type: ["string", "null"]
          format: uuid
        customerType:
          type: string
          enum: [guest, registered]
        preferredDate:
          type: ["string", "null"]
          format: date
        preferredTime:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/AdminRequestStatus"
        adminNotes:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        emailStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminRepairRequestUpdate:
      type: object
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/AdminRequestStatus"
        adminNotes:
          type: ["string", "null"]
          maxLength: 5000
        cancellationReason:
          type: ["string", "null"]
          maxLength: 2000
    AdminBookingStatus:
      type: string
      enum:
        - pending
        - contacted
        - confirmed
        - in_progress
        - completed
        - cancelled
    AdminBooking:
      type: object
      required:
        - bookingId
        - customerName
        - email
        - phone
        - customerId
        - customerType
        - applianceType
        - problemDescription
        - preferredDate
        - preferredTime
        - address
        - additionalNotes
        - status
        - adminNotes
        - cancellationReason
        - emailStatus
        - createdAt
        - updatedAt
      properties:
        bookingId:
          type: string
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        customerId:
          type: ["string", "null"]
          format: uuid
        customerType:
          type: string
          enum: [guest, registered]
        applianceType:
          type: string
        problemDescription:
          type: string
        preferredDate:
          type: string
          format: date
        preferredTime:
          type: string
        address:
          type: string
        additionalNotes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/AdminBookingStatus"
        adminNotes:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        emailStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminBookingUpdate:
      type: object
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/AdminBookingStatus"
        adminNotes:
          type: ["string", "null"]
          maxLength: 5000
        cancellationReason:
          type: ["string", "null"]
          maxLength: 2000
...